+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Master (unreleased)
 
+* Change default token type from "token" to "access"
+
 # v 0.12.0
 * Add `one_of` to permissions Plug to allow for OR'd sets of permissions as well as AND'd ones
 * Fix infinite recursion bug when joining channels